解决报错

+++ HARDWARE/src/gyroscope.c
@@ -26,7 +26,7 @@ uint8_t Check_Gyro_Data_Available(const uint8_t* gyro_raw_buf)
 void Parse_Gyro_Data(const uint8_t* gyro_raw_buf, Gyro_Wit_Data_t* gyro_data) {
     // 因为使用的是单轴陀螺仪，故只解析yaw轴
 	gyro_data->angle.roll_x = ((short)(gyro_raw_buf[3]<<8|gyro_raw_buf[2]))/32768.0*180;
-	gyro_data->angle.pitch_y = (gyro_raw_buf[5]<<8|gyro_raw_buf[4]))/32768.0*180;
+	gyro_data->angle.pitch_y = ((short)(gyro_raw_buf[5]<<8|gyro_raw_buf[4]))/32768.0*180;
     gyro_data->angle.yaw_z = ((short)(gyro_raw_buf[7] << 8 | gyro_raw_buf[6])) / 32768.0 * 180;
     if (gyro_data->angle.yaw_z < 0) {
         gyro_data->angle.yaw_z += 360;
